fix bug and add CORS

+++ app/Http/Controllers/AdminController.php
@@ -35,8 +35,8 @@ class AdminController extends Controller
         $user = Admin::create($validateData);
         $accessToken = $user->createToken('authToken')->accessToken;
 
-        return response()->json(['message'=>'account created successfully!','user'=>$user,'access_token'=>$accessToken],201);
-        
+        return response()->json(['message'=>'account created successfully!','user'=>$user,'access_token'=>$accessToken,'status_code'=>201],201);
+
     }
     public function login(Request $request)
     {
@@ -46,13 +46,13 @@ class AdminController extends Controller
         ]);
         $user = Admin::where('email',$loginData['email'])->first();
 
-        if(!$user || Hash::check($user->password, $request->password)){
-            return response(['message'=>'invalid credentials']);
+        if(!$user || !Hash::check($loginData['password'],$user->password)){
+            return response(['message'=>'invalid credentials','status_code'=>401],401);
         }
 
         $accessToken = $user->createToken('authToken')->accessToken;
 
-        return response()->json(['message'=>'log in successfully!','user'=>$user,'access_token'=>$accessToken],200);
+        return response()->json(['message'=>'log in successfully!','user'=>$user,'access_token'=>$accessToken,'status_code'=>200],200)->setStatusCode(200);
     }
     public function update(request $request){
         $validateData = $request->validate([
@@ -70,13 +70,17 @@ class AdminController extends Controller
         $data->password = Hash::make($request->password);
         $data->save();
 
-        return response()->json(["message" => "Data Updated Successfully!", "data" => $data],200);
+        return response()->json(["message" => "Data Updated Successfully!", "data" => $data,'status_code'=>200],200);
     }
 
     public function delete($id){
         $data = Admin::find($id);
-        $data->delete();
+        if($data){
+          $data->delete();
+        }else{
+          return response()->json(['status_code'=>400],400);
+        }
 
-        return response()->json(null,204);
+        return response()->json(['status_code'=>204],204);
     }
 }
