+++ app/Http/Controllers/TitikBanjirController.php
@@ -27,7 +27,7 @@ class TitikBanjirController extends Controller
             'nama_jalan',
             'kondisi_kerusakan',
             'foto',
-            'keterangan', 
+            'keterangan',
             DB::Raw('ST_AsGeoJSON(geometry) as geometry')
         )->where('id',$id)->first();;
     }
@@ -48,7 +48,7 @@ class TitikBanjirController extends Controller
 
         $data->geometry = json_decode($request->geometry);
 
-        return response()->json(["message" => "Data Added Successfully!", "data" => $data],201);
+        return response()->json(["message" => "Data Added Successfully!", "data" => $data,'status_code'=>201],201);
     }
 
     public function update(request $request, $id){
@@ -68,16 +68,20 @@ class TitikBanjirController extends Controller
         $data->foto = $request->foto;
         $data->keterangan = $request->$keterangan;
         $data->save();
-        
+
         $data->geometry = json_decode($request->geometry);
 
-        return response()->json(["message" => "Data Updated Successfully!", "data" => $data],200);
+        return response()->json(["message" => "Data Updated Successfully!", "data" => $data,'status_code'=>200],200);
     }
 
     public function delete($id){
         $data = TitikBanjir::find($id);
-        $data->delete();
+        if($data){
+          $data->delete();
+        }else{
+          return response()->json(['status_code'=>400],400);
+        }
 
-        return response()->json(null,204);
+        return response()->json(['status_code'=>204],204);
     }
 }
